+++ src/fr/eni/clinique/dal/ClientsDAOJdbcImpl.java
@@ -69,14 +69,14 @@ public class ClientsDAOJdbcImpl implements DaoClients {
         }
     }
 
-    public Clients selectById(int id) throws DALException {
+    public Clients selectById(Long id) throws DALException {
         try (            Connection connect = ConnectionSingleton.getConnect()
         ){
 
             String sql = "SELECT * FROM Clients WHERE CodeClient = ?";
             PreparedStatement stmt = connect.prepareStatement(sql);
 
-            stmt.setInt(1, id);//"reference,
+            stmt.setLong(1, id);//"reference,
 
             ResultSet res = stmt.executeQuery();
